+++ src/main/java/project/gsm/yaml/domain/humanities/entity/Awards.java
@@ -1,4 +1,4 @@
-package project.gsm.yaml.entity.major;
+package project.gsm.yaml.domain.humanities.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,19 +7,22 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
+@Table(name = "humanities_awards")
 @Entity
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Certificate {
+public class Awards {
     @Id
     @GeneratedValue
     private Long id;
+
     private String name;
-    private String startAt;
-    private String endAt;
+
+    private String fileUrl;
+
     @ManyToOne
-    @JoinColumn(name = "major")
-    private Major major;
+    @JoinColumn(name = "humanities")
+    private Humanities humanities;
 }
