+++ src/main/java/project/gsm/yaml/domain/humanities/entity/InnerSportsLeague.java
@@ -1,14 +1,14 @@
-package project.gsm.yaml.entity.humanities;
+package project.gsm.yaml.domain.humanities.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum Marathon {
-    HALF(50),
-    TEN(30),
-    FIVE(20),
+public enum InnerSportsLeague {
+    THREE(50),
+    TWO(30),
+    ONE(20),
     NONE(0);
     private Integer score;
 }
