+++ src/main/java/project/gsm/yaml/domain/humanities/repository/AwardsRepository.java
@@ -1,0 +1,12 @@
+package project.gsm.yaml.domain.humanities.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import project.gsm.yaml.domain.humanities.entity.Awards;
+
+import java.util.Optional;
+
+@Repository
+public interface AwardsRepository extends JpaRepository<Awards, Long> {
+
+}
