+++ src/main/java/project/gsm/yaml/domain/humanities/repository/HumanitiesRepository.java
@@ -1,0 +1,9 @@
+package project.gsm.yaml.domain.humanities.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import project.gsm.yaml.domain.humanities.entity.Humanities;
+
+@Repository
+public interface HumanitiesRepository extends JpaRepository<Humanities, Long> {
+}
