+++ src/main/java/project/gsm/yaml/domain/humanities/service/HumanitiesService.java
@@ -1,0 +1,13 @@
+package project.gsm.yaml.domain.humanities.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import project.gsm.yaml.domain.humanities.repository.HumanitiesRepository;
+
+@Service
+@RequiredArgsConstructor
+public class HumanitiesService {
+
+    private final HumanitiesRepository humanitiesRepository;
+
+}
