+++ src/main/java/project/gsm/yaml/domain/major/Major.java
@@ -1,4 +1,4 @@
-package project.gsm.yaml.entity.major;
+package project.gsm.yaml.domain.major;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
