+++ src/main/java/project/gsm/yaml/domain/major/Prize.java
@@ -1,4 +1,4 @@
-package project.gsm.yaml.entity.major;
+package project.gsm.yaml.domain.major;
 
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
