--- src/main/java/project/gsm/yaml/repository/MajorRepository.java
@@ -1,9 +1,0 @@
-package project.gsm.yaml.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import project.gsm.yaml.entity.major.Major;
-
-@Repository
-public interface MajorRepository extends JpaRepository<Major, Long> {
-}
